Adds flags to disable the purge of queues on AbstractJmsFunctionalTestCase. Change required for new tests of EE-1999
git-svn-id: https://svn.codehaus.org/mule/branches/mule-2.2.x@19284 bf997673-6b11-0410-b953-e057580c5b09

+++ transports/jms/src/test/java/org/mule/transport/jms/integration/AbstractJmsFunctionalTestCase.java
@@ -121,6 +121,8 @@ public abstract class AbstractJmsFunctionalTestCase extends FunctionalTestCase
     protected Scenario scenarioRollback;
     protected Scenario scenarioNotReceive;
     protected Scenario scenarioReceive;
+    protected boolean purgeQueuesOnPreSetUp = true;
+    protected boolean purgeQueuesOnTearDown = true;
 
     private MuleClient client = null;
 
@@ -225,9 +227,12 @@ public abstract class AbstractJmsFunctionalTestCase extends FunctionalTestCase
     {
         super.suitePreSetUp();
 
-        purge(getInboundQueueName());
-        purge(getOutboundQueueName());
-        // TODO Add other queues if necessary
+        if (purgeQueuesOnPreSetUp)
+        {
+            purge(getInboundQueueName());
+            purge(getOutboundQueueName());
+            // TODO Add other queues if necessary
+        }
     }
     
     /**
@@ -510,9 +515,12 @@ public abstract class AbstractJmsFunctionalTestCase extends FunctionalTestCase
 
     protected void doTearDown() throws Exception
     {
-        purge(getInboundQueueName());
-        purge(getOutboundQueueName());
-        purgeTopics();
+        if (purgeQueuesOnTearDown)
+        {
+            purge(getInboundQueueName());
+            purge(getOutboundQueueName());
+            purgeTopics();
+        }
 
         super.doTearDown();
         if (client != null)
